v0.4
- Postprocessing method completed
- Needs calibration and further improvements
- While loops have been added to the SD card functions
- Timeout for while loops needs to be implemented

+++ ByteFunctions.h
@@ -35,6 +35,6 @@ unsigned char DetermineNumberOfTriggeredChannels(unsigned char* Array, unsigned 
 
 void UlToStr(char *s, unsigned long bin, unsigned char n);
 
-double ConvertToEnergyAnode(unsigned int ADCValue);
-double ConvertToEnergyCathode(unsigned int ADCValue);
+double ConvertToEnergyAnode(double ADCValue);
+double ConvertToEnergyCathode(double ADCValue);
 #endif /* BYTEFUNCTIONS_H_ */
